+++ packages/mantine/src/components/inline-confirm/index.ts
@@ -1,2 +1,0 @@
-export * from './InlineConfirm.js';
-export * from './InlineConfirmContext.js';