+++ packages/mantine/src/components/modal/Modal.tsx
@@ -4,7 +4,7 @@ import {
     ModalFactory as MantineModalFactory,
     ModalProps as MantineModalProps,
 } from '@mantine/core';
-import {Header, HeaderDocAnchorProps, HeaderProps} from '../header/index.js';
+import {Header, type HeaderDocAnchorProps, type HeaderProps} from '../header/Header.js';
 import {ModalFooter as PlasmaModalFooter} from './ModalFooter.js';
 
 interface PlasmaModalProps extends MantineModalProps {