+++ packages/mantine/src/components/modal/index.ts
@@ -1,2 +1,0 @@
-export * from './Modal.js';
-export * from './ModalFooter.js';