+++ packages/mantine/src/components/table/layouts/row-layout/RowLayoutHeader.tsx
@@ -1,6 +1,6 @@
 import {BoxProps, CompoundStylesApiProps, Factory, useProps} from '@mantine/core';
 import {ForwardedRef} from 'react';
-import {CustomComponentThemeExtend, identity} from '../../../../utils/index.js';
+import {CustomComponentThemeExtend, identity} from '../../../../utils/createFactoryComponent.js';
 import {TableLayoutProps} from '../../Table.types.js';
 import {useTableContext} from '../../TableContext.js';
 import {Th} from '../../table-header/Th.js';