+++ packages/mantine/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useParentHeight.js';
-export * from './useControlledList.js';