+++ packages/mantine/src/theme/plasmaCSSVariablesResolver.ts
@@ -31,7 +31,7 @@ export const plasmaCSSVariablesResolver: CSSVariablesResolver = (theme) => {
 
             // mantine overrides
             '--mantine-primary-color-light': 'var(--mantine-primary-color-7)',
-            '--mantine-color-body': theme.colors.dark[8],
+            '--mantine-color-body': theme.colors.dark[7],
             '--mantine-color-default': theme.colors.dark[7],
             '--mantine-color-default-border': theme.colors.dark[6],
             '--mantine-color-default-hover': theme.colors.dark[6],
@@ -56,7 +56,7 @@ export const plasmaCSSVariablesResolver: CSSVariablesResolver = (theme) => {
             '--coveo-color-bg-dark-surface': theme.colors.violet[9],
 
             // mantine overrides
-            '--mantine-color-body': theme.colors.gray[0],
+            '--mantine-color-body': 'var(--mantine-color-white)',
             '--mantine-color-default': 'var(--mantine-color-white)',
             '--mantine-color-default-border': theme.colors.gray[2],
             '--mantine-color-default-hover': theme.colors.gray[1],